feat(frontend):alignment

+++ frontend/src/api.rs
@@ -77,7 +77,7 @@ pub struct Book {
     #[serde(rename = "coverUrl", default)]
     pub cover_url: String,
     #[serde(default)]
-    pub stock: i32,
+    pub stock: i64,
     #[serde(rename = "createdAt", default)]
     pub created_at: String,
     #[serde(rename = "updatedAt", default)]
@@ -93,7 +93,7 @@ pub struct BookCreate {
     #[serde(rename = "coverKey", default, skip_serializing_if = "String::is_empty")]
     pub cover_key: String,
     #[serde(default)]
-    pub stock: i32,
+    pub stock: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq)]
@@ -107,7 +107,7 @@ pub struct BookUpdate {
     #[serde(rename = "coverKey", default, skip_serializing_if = "String::is_empty")]
     pub cover_key: String,
     #[serde(default)]
-    pub stock: i32,
+    pub stock: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq)]
@@ -248,7 +248,11 @@ async fn parse_envelope<T: for<'de> Deserialize<'de>>(
         .map_err(|e| ApiError::Network(e.to_string()))?;
     if !(200..300).contains(&status) {
         if let Ok(env) = serde_json::from_str::<ApiResponse<serde_json::Value>>(&text) {
-            let code = if env.code != 0 { env.code } else { status as i32 };
+            let code = if env.code != 0 {
+                env.code
+            } else {
+                status as i32
+            };
             return Err(ApiError::Message {
                 code,
                 msg: if env.msg.is_empty() {
@@ -275,8 +279,7 @@ async fn parse_envelope<T: for<'de> Deserialize<'de>>(
             },
         });
     }
-    env.data
-        .ok_or_else(|| ApiError::msg("empty response data"))
+    env.data.ok_or_else(|| ApiError::msg("empty response data"))
 }
 
 async fn parse_empty(resp: gloo_net::http::Response) -> ApiResult<()> {
@@ -287,7 +290,11 @@ async fn parse_empty(resp: gloo_net::http::Response) -> ApiResult<()> {
         .map_err(|e| ApiError::Network(e.to_string()))?;
     if !(200..300).contains(&status) {
         if let Ok(env) = serde_json::from_str::<ApiResponse<serde_json::Value>>(&text) {
-            let code = if env.code != 0 { env.code } else { status as i32 };
+            let code = if env.code != 0 {
+                env.code
+            } else {
+                status as i32
+            };
             return Err(ApiError::Message {
                 code,
                 msg: if env.msg.is_empty() {
@@ -461,8 +468,7 @@ pub async fn get_book(id: i64) -> ApiResult<Book> {
 }
 
 pub async fn create_book(input: BookCreate) -> ApiResult<IdOnly> {
-    let body =
-        serde_json::to_string(&input).map_err(|e| ApiError::msg(format!("encode: {e}")))?;
+    let body = serde_json::to_string(&input).map_err(|e| ApiError::msg(format!("encode: {e}")))?;
     let url = format!("{}/books", API_BASE);
     let body = body.clone();
     let resp = send_guarded(move || {
@@ -475,8 +481,7 @@ pub async fn create_book(input: BookCreate) -> ApiResult<IdOnly> {
 }
 
 pub async fn update_book(id: i64, input: BookUpdate) -> ApiResult<()> {
-    let body =
-        serde_json::to_string(&input).map_err(|e| ApiError::msg(format!("encode: {e}")))?;
+    let body = serde_json::to_string(&input).map_err(|e| ApiError::msg(format!("encode: {e}")))?;
     let url = format!("{}/books/{id}", API_BASE);
     let body = body.clone();
     let resp = send_guarded(move || {
@@ -513,8 +518,7 @@ pub async fn list_loans(offset: i32, limit: i32) -> ApiResult<Vec<LoanRecord>> {
 }
 
 pub async fn create_loan(input: LoanCreate) -> ApiResult<IdOnly> {
-    let body =
-        serde_json::to_string(&input).map_err(|e| ApiError::msg(format!("encode: {e}")))?;
+    let body = serde_json::to_string(&input).map_err(|e| ApiError::msg(format!("encode: {e}")))?;
     let url = format!("{}/loans", API_BASE);
     let body = body.clone();
     let resp = send_guarded(move || {
@@ -533,8 +537,7 @@ pub async fn get_loan(id: i64) -> ApiResult<LoanRecord> {
 }
 
 pub async fn update_loan(id: i64, input: LoanUpdate) -> ApiResult<()> {
-    let body =
-        serde_json::to_string(&input).map_err(|e| ApiError::msg(format!("encode: {e}")))?;
+    let body = serde_json::to_string(&input).map_err(|e| ApiError::msg(format!("encode: {e}")))?;
     let url = format!("{}/loans/{id}", API_BASE);
     let body = body.clone();
     let resp = send_guarded(move || {
